+++ src/checkout/components/index.ts
@@ -1,4 +1,3 @@
-export { default as AddressSummary } from "./AddressSummary";
 export { default as ShippingOptionSummary } from "./ShippingOptionSummary";
 export { default as StepCheck } from "./StepCheck";
 export { default as Steps } from "./Steps";