+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable, unreleased changes to this project will be documented in this file.
 
 - Add stock quantity check without checkout in cart page - #254 by @piotrgrundas
 - Add ability to chose payment method, add dummy payment method, improve error handling on checkout shipping address update - #255 by @piotrgrundas
+- Add order details page - #262 by @piotrgrundas
 
 ## 0.4.0